requirements.txt
fitz==0.0.1.dev2
pdfplumber==0.8.0
Pillow==9.1.0
reportlab==3.6.2